Add update command for lock and template
