+++ lib/wombat/deploy.rb
@@ -17,7 +17,7 @@ class DeployRunner
     case cloud
     when 'aws'
       update_lock(cloud) if lock_opt
-      create_template if template_opt
+      update_template(cloud) if template_opt
       create_stack(stack)
     end
   end
@@ -44,69 +44,4 @@ class DeployRunner
     puts "Created: #{resp.stack_id}"
   end
 
-  def create_template
-    region = lock['aws']['region']
-    @chef_server_ami = lock['amis'][region]['chef-server']
-    @automate_ami = lock['amis'][region]['automate']
-    @compliance_ami = lock['amis'][region]['compliance']
-    @build_nodes = lock['build-nodes']['count'].to_i
-    @build_node_ami = {}
-    1.upto(@build_nodes) do |i|
-      @build_node_ami[i] = lock['amis'][region]['build-node'][i.to_s]
-    end
-    @infra = {}
-    infranodes.each do |name, _rl|
-      @infra[name] = lock['amis'][region]['infranodes'][name]
-    end
-    @workstations = lock['workstations']['count'].to_i
-    @workstation_ami = {}
-    1.upto(@workstations) do |i|
-      @workstation_ami[i] = lock['amis'][region]['workstation'][i.to_s]
-    end
-    @availability_zone = lock['aws']['az']
-    @iam_roles = lock['aws']['iam_roles']
-    @demo = lock['name']
-    @version = lock['version']
-    @ttl = lock['ttl']
-    rendered_cfn = ERB.new(File.read("#{conf['template_dir']}/cfn.json.erb"), nil, '-').result(binding)
-    File.open("#{conf['stack_dir']}/#{@demo}.json", 'w') { |file| file.puts rendered_cfn }
-    banner("Generated: #{conf['stack_dir']}/#{@demo}.json")
-  end
-
-  def logs
-    Dir.glob("#{conf['log_dir']}/#{cloud}*.log").reject { |l| !l.match(wombat['linux']) }
-  end
-
-  def update_lock(cloud)
-    banner('Updating wombat.lock')
-
-    copy = {}
-    copy = wombat
-    region = copy[cloud]['region']
-    linux = copy['linux']
-    copy['amis'] = { region => {} }
-    logs.each do |log|
-      case log
-      when /build-node/
-        copy['amis'][region]['build-node'] ||= {}
-        num = log.split('-')[3]
-        copy['amis'][region]['build-node'].store(num, parse_log(log, cloud))
-      when /workstation/
-        copy['amis'][region]['workstation'] ||= {}
-        num = log.split('-')[2]
-        copy['amis'][region]['workstation'].store(num, parse_log(log, cloud))
-      when /infranodes/
-        copy['amis'][region]['infranodes'] ||= {}
-        name = log.split('-')[2]
-        copy['amis'][region]['infranodes'].store(name, parse_log(log, cloud))
-      else
-        instance = log.match("#{cloud}-(.*)-(.*)\.log")[1]
-        copy['amis'][region].store(instance, parse_log(log, cloud))
-      end
-    end
-    copy['last_updated'] = Time.now.gmtime.strftime('%Y%m%d%H%M%S')
-    File.open('wombat.lock', 'w') do |f|
-      f.write(JSON.pretty_generate(copy))
-    end
-  end
 end
